backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Dmitri Khassanov <khdmitri@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
alembic = "^1.11.3"
fastapi-sqlalchemy = "^0.2.1"
pydantic = "^2.3.0"
psycopg2 = "^2.9.7"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
bump-pydantic = "^0.6.1"
pydantic-settings = "^2.0.3"
jwt = "^1.3.1"
python-jose = "^3.3.0"
passlib = "1.7.4"
emails = "^0.6"
email-validator = "^2.0.0.post2"
asyncpg = "^0.28.0"
python-multipart = "^0.0.6"
arq = "^0.25.0"
httpx = "^0.24.1"
aiogram = "^3.0.0"
facebook-sdk = "^3.1.0"
requests = "^2.31.0"
asyncer = "^0.0.2"
vk-api = "^11.9.9"
chatgpt = "^2.2212.0"
browser-cookie3 = "^0.19.1"
pycryptodome = "^3.19.0"
pyexecjs = "^1.5.1"
curl-cffi = "^0.5.9"
nest-asyncio = "^1.5.8"
watchfiles = "^0.21.0"
py-bcrypt = "^0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
